# Problem: common parameters not cached