--- a/x/evm/keeper/keeper.go
+++ b/x/evm/keeper/keeper.go
@@ -199,8 +199,8 @@ func (k *Keeper) PostTxProcessing(ctx sdk.Context, msg core.Message, receipt *et
 }
 
 // Tracer return a default vm.Tracer based on current keeper state
-func (k Keeper) Tracer(ctx sdk.Context, msg core.Message, ethCfg *params.ChainConfig) vm.EVMLogger {
-	return types.NewTracer(k.tracer, msg, ethCfg, ctx.BlockHeight(), ctx.BlockHeader().Time.Unix())
+func (k Keeper) Tracer(msg core.Message, rules params.Rules) vm.EVMLogger {
+	return types.NewTracer(k.tracer, msg, rules)
 }
 
 // GetAccount load nonce and codehash without balance,