--- a/x/evm/keeper/params.go
+++ b/x/evm/keeper/params.go
@@ -23,20 +23,13 @@ import (
 
 // GetParams returns the total set of evm parameters.
 func (k Keeper) GetParams(ctx sdk.Context) types.Params {
-	var params types.Params
-	objStore := ctx.ObjectStore(k.objectKey)
-	v := objStore.Get(types.KeyPrefixObjectParams)
-	if v == nil {
-		bz := ctx.KVStore(k.storeKey).Get(types.KeyPrefixParams)
-		if len(bz) != 0 {
-			k.cdc.MustUnmarshal(bz, &params)
-		} else {
-			params = k.GetLegacyParams(ctx)
-		}
-		objStore.Set(types.KeyPrefixObjectParams, &params)
-	} else {
-		params = *(v.(*types.Params))
+	bz := ctx.KVStore(k.storeKey).Get(types.KeyPrefixParams)
+	if len(bz) == 0 {
+		return k.GetLegacyParams(ctx)
 	}
+
+	var params types.Params
+	k.cdc.MustUnmarshal(bz, &params)
 	return params
 }
 
@@ -48,12 +41,6 @@ func (k Keeper) SetParams(ctx sdk.Context, p types.Params) error {
 	store := ctx.KVStore(k.storeKey)
 	bz := k.cdc.MustMarshal(&p)
 	store.Set(types.KeyPrefixParams, bz)
-
-	// set to cache as well, decode again to be compatible with the previous behavior
-	var params types.Params
-	k.cdc.MustUnmarshal(bz, &params)
-	ctx.ObjectStore(k.objectKey).Set(types.KeyPrefixObjectParams, &params)
-
 	return nil
 }