--- a/x/evm/types/tracer.go
+++ b/x/evm/types/tracer.go
@@ -36,7 +36,7 @@ const (
 
 // NewTracer creates a new Logger tracer to collect execution traces from an
 // EVM transaction.
-func NewTracer(tracer string, msg core.Message, cfg *params.ChainConfig, height, time int64) vm.EVMLogger {
+func NewTracer(tracer string, msg core.Message, rules params.Rules) vm.EVMLogger {
 	// TODO: enable additional log configuration
 	logCfg := &logger.Config{
 		Debug: true,
@@ -44,7 +44,7 @@ func NewTracer(tracer string, msg core.Message, cfg *params.ChainConfig, height,
 
 	switch tracer {
 	case TracerAccessList:
-		preCompiles := vm.DefaultActivePrecompiles(cfg.Rules(big.NewInt(height), cfg.MergeNetsplitBlock != nil, uint64(time)))
+		preCompiles := vm.DefaultActivePrecompiles(rules)
 		return logger.NewAccessListTracer(msg.AccessList, msg.From, *msg.To, preCompiles)
 	case TracerJSON:
 		return logger.NewJSONLogger(logCfg, os.Stderr)